requirements.txt
cryptography~=42.0.7
pip~=24.0
redis~=5.0.5
amqp~=5.2.0
vine~=5.1.0
setuptools~=56.0.0
pytz~=2024.1
Django~=4.2.13
kombu~=5.3.7
billiard~=4.2.0
requests~=2.32.2
celery~=5.4.0
click~=8.1.7
python-dateutil~=2.9.0.post0
asgiref~=3.8.1
sqlparse~=0.5.0
colorama~=0.4.6
django-allauth~=0.61.1
PyJWT~=2.8.0
certifi~=2024.2.2
tzlocal~=5.2
urllib3~=2.2.1
wcwidth~=0.2.13
cffi~=1.16.0
six~=1.16.0
Markdown~=3.6
oauthlib~=3.2.2
idna~=3.7
pycparser~=2.22
APScheduler~=3.10.4
django-filter~=24.2
djangorestframework~=3.15.1
django-modeltranslation~=0.19.4
zipp~=3.19.2
python-dotenv~=1.0.1